After working comments final
- Add VPM
- Clear code
- Hierarchy folders split
